==> partII_CAMs/02_partII_CAMs.qmd <==
---
title: "Analyze CAM data on micro, meso, macro level for part II of basal attributes article"
author: "Paul Sölder, Julius Fenn"
format:
  html:
    toc: true
    toc-depth: 3
    html-math-method: katex
    number-sections: true
---


# Notes


# load data files



```{r}
#| label: load files
#| warning: false

# sets the directory of location of this script as the current directory
# setwd(dirname(rstudioapi::getSourceEditorContext()$path))

### load packages
require(pacman)
p_load('tidyverse', 'jsonlite', 'magrittr', 'xlsx',
       'stargazer', 'psych', 'jtools', 'DT', 'ggstatsplot', 
       'lavaan', 
       'regsem', 'MplusAutomation', 'igraph', 'shiny', 'ggplot2', 'tidyLPA', 'MultilayerExtraction',
       'Matrix', 'igraph', 'foreach', 'doParallel', 'parallel', 'R.matlab')
 

multilayer_extraction_results <- readRDS(file = "outputs/multilayer_extraction_results.Rda")

setwd("outputs/01_dataPreperation/final")
### load questionnaire
questionnaire <- readRDS(file = "questionnaire.rds")

CAMfiles <- readRDS(file = "CAMfiles.rds")

CAMdrawn <- readRDS(file = "CAMdrawn.rds")

CAMaggregated <- readRDS(file = "CAMaggregated.rds")

networkIndicators <- readRDS(file = "networkIndicators.rds")

CAMwordlist <- xlsx::read.xlsx2(file = "CAMwordlist.xlsx", sheetIndex = 1)


### load functions
# print(getwd())
setwd("../../../functions")
for(i in 1:length(dir())){
  # print(dir()[i])
  source(dir()[i], encoding = "utf-8")
}


setwd("../functions_CAMapp")
for(i in 1:length(dir())){
  # print(dir()[i])
  source(dir()[i], encoding = "utf-8")
}
rm(i)



### summary function
data_summary <- function(data, varname, groupnames){
  require(plyr)
  summary_func <- function(x, col){
    c(mean = mean(x[[col]], na.rm=TRUE),
      se = sd(x[[col]], na.rm=TRUE) / sqrt(length(x[[col]])))
  }
  data_sum<-ddply(data, groupnames, .fun=summary_func,
                  varname)
  data_sum <- plyr::rename(data_sum, c("mean" = varname))
  return(data_sum)
}

```



# Micro level

Distributions, cluster analyses... on single node level...

## Valence distributions

```{r}
#| label: valence distributions

valenceData <- CAMfiles[[1]]
valenceData$value[valenceData$value == 10] <- 0

valenceDataShort <- valenceData %>% pivot_wider(names_from = text, values_from = value, id_cols = CAM)

hist1_names <- c("Akzeptanz eines neuen Materialsystems", "Energie   generierend", 
"Energie   speichernd", "Insekten ähnlich", "aktive Formänderung durch Umwelteinwirkung", 
"aktive Verhaltensänderung durch Umwelteinwirkung", "autonom", 
"bioinspiriert", "elektronikfrei", "energieautonom", "energieeffizient", 
"enthält Kunststoff", "haltbar", "intelligent", "langlebig", 
"lebensähnlich")
hist2_names <- c("leicht zerstörbar", "multifunktional", "nachhaltig", 
"passive Formänderung durch Umwelteinwirkung", "passive Verhaltensänderung durch Umwelteinwirkung", 
"reaktionsfähig", "robust", "selbstheilend", "selbstreparierend", 
"technologisch", "umweltfreundlich", "umweltschädlich", "wartungsfrei", 
"wartungsintensiv", "widerstandsfähig", "zuverlässig", "ökologisch")


# adjust aesthetics of ggplot
ggplot_theme <- theme(axis.title.x = element_blank(),
                      axis.title.y = element_text(size=12),
                      axis.text.x = element_text(size=10,hjust=0.5,vjust=0.5,face="plain", colour = "black"),
                      axis.text.y = element_text(size=12,face="plain", colour = "black"),
                      panel.border = element_blank(),
                      axis.line = element_line(colour = "black"),
                      panel.grid.major = element_blank(),
                      panel.grid.minor = element_blank(),
                      panel.background = element_blank())

valenceData %>% filter(text %in% hist1_names) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~text, ncol = 4, nrow = 4) + ggplot_theme
valenceData %>% filter(text %in% hist2_names) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~text, ncol = 4, nrow = 5)


# histograms sorted by MEAN VALENCE
hist1_names_mean <- CAMwordlist[order(CAMwordlist$mean_valence, decreasing = TRUE), "Words"][1:16]
hist2_names_mean <- CAMwordlist[order(CAMwordlist$mean_valence, decreasing = TRUE), "Words"][17:33]

valenceData %>% filter(text %in% hist1_names_mean) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~ factor(text, levels = hist1_names_mean), ncol = 4, nrow = 4) +
  ggtitle("Sorted by avg. valence")
valenceData %>% filter(text %in% hist2_names_mean) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~ factor(text, levels = hist2_names_mean), ncol = 4, nrow = 5) +
  ggtitle("Sorted by avg. valence")


# histograms sorted by MEAN DEGREE
hist1_names_degree <- CAMwordlist[order(CAMwordlist$mean_degree, decreasing = TRUE), "Words"][1:16]
hist2_names_degree <- CAMwordlist[order(CAMwordlist$mean_degree, decreasing = TRUE), "Words"][17:33]

valenceData %>% filter(text %in% hist1_names_degree) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~ factor(text, levels = hist1_names_degree), ncol = 4, nrow = 4) +
  ggtitle("Sorted by avg. degree")
valenceData %>% filter(text %in% hist2_names_degree) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~ factor(text, levels = hist2_names_degree), ncol = 4, nrow = 5) +
  ggtitle("Sorted by avg. degree")


# histograms sorted by MEAN TRANSITIVITY
hist1_names_transitivity <- CAMwordlist[order(CAMwordlist$mean_transitivity, decreasing = TRUE), "Words"][1:16]
hist2_names_transitivity <- CAMwordlist[order(CAMwordlist$mean_transitivity, decreasing = TRUE), "Words"][17:33]

valenceData %>% filter(text %in% hist1_names_transitivity) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~ factor(text, levels = hist1_names_transitivity), ncol = 4, nrow = 4) +
  ggtitle("Sorted by avg. transitivity")
valenceData %>% filter(text %in% hist2_names_transitivity) %>%
  ggplot(aes(x = value)) + geom_histogram() + facet_wrap(~ factor(text, levels = hist2_names_transitivity), ncol = 4, nrow = 5) +
  ggtitle("Sorted by avg. transitivity")
```



# Meso level

## Multilayer community detection

### MultilayerExtraction
Community detection with the MultilayerExtraction-Algorithm (https://github.com/jdwilson4/MultilayerExtraction)

```{r}
setwd("outputs")
# recode vertex labels to integers
attributes_names <- c(hist1_names, hist2_names)
CAMaggregated[[4]][["node1"]] <- match(CAMaggregated[[4]][["node1"]], attributes_names)
CAMaggregated[[4]][["node2"]] <- match(CAMaggregated[[4]][["node2"]], attributes_names)

# recode layers (CAM IDs) to integers
layer_names <- unique(CAMfiles[[1]]$participantCAM)
CAMaggregated[[4]][["layer"]] <- match(CAMaggregated[[4]][["layer"]], layer_names)

# now we're ready for community extraction using the MultiLayer extraction package
set.seed(123)
start_time <- Sys.time()
#multilayer_extraction_results <- multilayer.extraction(adjacency = CAMaggregated[[4]], seed = 123, min.score = 0, prop.sample = .10)
end_time <- Sys.time()
end_time - start_time

#saveRDS(multilayer_extraction_results, file = "multilayer_extraction_results.Rda")

plot(multilayer_extraction_results, main = "Diagnostic Plot")


# look at the first ... 20 communities
multilayer_extraction_communities <- data.frame(n_communities = seq(1:20))
multilayer_extraction_communities[paste0("community", 1:20)] <- NA

for(i in 1:20) {
  object <- refine(multilayer_extraction_results, k = i, m = 192, n = nrow(CAMwordlist))
  
  for(j in 1:i) {
    multilayer_extraction_communities[i,j+1] <- paste(attributes_names[which(object$Vertices[,j] == 1)], collapse = ", ")
  }
}

DT::datatable(multilayer_extraction_communities, options = list(pageLength = 5)) 
```


### GenLouvain
Community detection with the generalized Louvain algorithm as proposed by Mucha et al. (2010).

Re-import the results from GenLouvain as processed by MATLAB.

```{r}
genlouvain <- readMat("outputs/genlouvain_results.mat", drop = "singletonLists")
genlouvain <- unlist(genlouvain, recursive = FALSE)

genlouvain_results <- list()

for(i in 1:length(genlouvain)) {
  tmp_result <- unlist(unlist(genlouvain[[i]], recursive = FALSE)[[1]], recursive = FALSE)
  tmp_result <- matrix(data = tmp_result, nrow = nrow(CAMwordlist), ncol = length(CAMaggregated[[5]]))
  rownames(tmp_result) <- rownames(CAMaggregated[[5]][[1]])
  colnames(tmp_result) <- names(CAMaggregated[[5]])
  
  # named list with parameter 'omega' as name and attribute
  attributes(tmp_result)$omega <- unlist(unlist(genlouvain[[i]], recursive = FALSE)[[4]], recursive = FALSE)
  genlouvain_results[[paste0("omega", unlist(unlist(genlouvain[[i]], recursive = FALSE)[[4]], recursive = FALSE))]] <- 
    tmp_result
}
```

Number of communities found:
```{r}
for(i in 1:length(genlouvain_results)) {
  print(paste("Found", length(unique(as.vector(genlouvain_results[[i]]))), "communities with ",
              names(genlouvain_results)[i]))
}
```

Take a look at the communities:

```{r}

genlouvain_communities <- data.frame(omega = seq(0.1, 1, 0.1), community1 = NA, community2 = NA, community3 = NA,
                                     community4 = NA, community5 = NA)

# omega = 0 with no coupling of the slices is not interesting, there are >900 communities.
# therefore we start at index 2 (omega = 0.1)
genlouvain_results <- genlouvain_results[-1]

for(i in 1:length(genlouvain_results)) {
  for(j in 1:max(unlist(genlouvain_results))) {
    genlouvain_communities[i, j+1] <- paste(rownames(genlouvain_results[[i]])[genlouvain_results[[i]][,1] == j], collapse=", ")
  }
}

DT::datatable(genlouvain_communities, options = list(pageLength = 5)) 

```
	
	
## random graph - delete me

Compare the probability that two concepts are connected in randomly generated networks to drawn CAMs:

```{r}
## get average number of drawn concepts (here fixed)
numConcepts <- mean(networkIndicators$num_nodes_macro)
## get average density
numDensity <- mean(networkIndicators$density_macro)

## simply get the average probability that two concepts are connected:
g <- igraph::random.graph.game(n = numConcepts, p.or.m = numDensity)
plot(g)
are.connected(g, 1, 2)


#> whereby each edge to be drawn has the identical probability in the Erdős–Rényi model
vec_booleanConnected <- c()
# vec_booleanConnected2 <- c()

for(i in 1:10000){
  g <- igraph::random.graph.game(n = numConcepts, p.or.m = numDensity)
  vec_booleanConnected[i] <- are.connected(g, 1, 2)
    # vec_booleanConnected2[i] <- are.connected(g, 22, 26)
}
baselineProbability <- mean(vec_booleanConnected)
baselineProbability
# mean(vec_booleanConnected2)
```


This baseline probability can be compared to all possible combinations of drawn concepts: 

```{r}
vec_boolean <- c()
for(i in 1:length(CAMdrawn)){
  vec_boolean[i] <- are.connected(graph = CAMdrawn[[i]], 
              v1 = V(CAMdrawn[[i]])[V(CAMdrawn[[i]])$label == "robust"], 
              v2 = V(CAMdrawn[[i]])[V(CAMdrawn[[i]])$label == "widerstandsfähig"])
}

mean(vec_boolean)
```







# Macro level

## plot aggregated CAM

```{r}
#| label: aggregate CAMs

# plot(CAMaggregated[[2]], vertex.size=diag(CAMaggregated[[1]]) / max(diag(CAMaggregated[[1]]))*20, edge.arrow.size=0.01)
# plot(CAMaggregated[[2]], vertex.size=(abs(V(CAMaggregated[[2]])$value)+1)*5, edge.arrow.size=0.01)


g = CAMaggregated[[2]]
g2 = simplify(CAMaggregated[[2]])
# plot(g2, edge.arrow.size=0.01,
#      vertex.size=diag(CAMaggregated[[1]]) / max(diag(CAMaggregated[[1]]))*20)

E(g2)$weight = sapply(E(g2), function(e) {
  length(all_shortest_paths(g, from=ends(g2, e)[1], to=ends(g2, e)[2])$res) } )
E(g2)$weight = E(g2)$weight * 2
# E(g2)$weight[E(g2)$weight == 1] <- NA

V(g2)$color[V(g2)$value <= .5 & V(g2)$value >= -.5] <- "yellow"

V(g2)$shape <- NA
V(g2)$shape <- ifelse(test = V(g2)$color == "yellow", yes = "square", no = "circle")



### > plot multiple times because of random layout
for(i in 1:3){
  plot(g2, edge.arrow.size = 0,
     layout=layout_nicely, vertex.frame.color="black", asp = .5, margin = -0.1,
     vertex.size=diag(CAMaggregated[[1]]) / max(diag(CAMaggregated[[1]]))*5,
     vertex.label.cex = .9, 
     edge.weight=2, edge.width=(E(g2)$weight/35))
}



CAMaggregated[[1]][1:5, 1:5]
```





# Clusteranalyses



## hierachical cluster analysis


```{r}
hc_dat <- data.frame(CAM = unique(CAMfiles[[1]]$CAM), participantCAM = unique(CAMfiles[[1]]$participantCAM))

### create data set
## word vars
for(w in unique(CAMfiles[[1]]$text)){
  varName_w <- str_remove_all(string = str_to_title(string = w, locale = "en"), pattern = " |\\W+")

  hc_dat[[paste0("N_", varName_w)]] <- NA
  hc_dat[[paste0("mean_", varName_w)]] <- NA
  hc_dat[[paste0("SD_", varName_w)]] <- NA
}


## get N, mean, sd of single summarized concepts
verbose = FALSE

for(c in unique(CAMfiles[[1]]$CAM)){
  if(verbose){
      cat("considered CAM: ", c, "\n")
  }
  tmp_CAM_nodes <- CAMfiles[[1]][CAMfiles[[1]]$CAM == c, ]
  tmp_CAM_nodes$value <- ifelse(test = tmp_CAM_nodes$value == 10, yes = 0, no = tmp_CAM_nodes$value)

  for(w in unique(CAMfiles[[1]]$text)){
      if(verbose){
    cat("considered concept: ", w, "\n")
      }

    varName_w <- str_remove_all(string = str_to_title(string = w, locale = "en"), pattern = " |\\W+")
          if(verbose){
    cat("   > the freqeuncy, mean, SD are saved with the prefix N_, mean_, SD_ plus
      word without white spaces: ", varName_w, "\n")
          }

    if(sum(tmp_CAM_nodes$text == w) > 0){
      tmp_CAM_nodes_w <- tmp_CAM_nodes[tmp_CAM_nodes$text == w, ]

      ## add N
      hc_dat[hc_dat$CAM == c, paste0("N_", varName_w)] <- nrow(tmp_CAM_nodes_w)
      ## add mean
      hc_dat[hc_dat$CAM == c, paste0("mean_", varName_w)] <- mean(x = tmp_CAM_nodes_w$value)
      ## add SD, only if > 1
      hc_dat[hc_dat$CAM == c, paste0("SD_", varName_w)] <- sd(x = tmp_CAM_nodes_w$value)
    }
  }
        if(verbose){
  cat("\n")
        }
}


### compute clustering
hc_df <- hc_dat[, str_subset(string = colnames(hc_dat), pattern = "mean_")]
hc_df <- hc_df[, colSums(x = !is.na(hc_df)) >= 2] # only considers concepts which where drawn at least 2 times


hc_df_scaled <- scale(hc_df)

dist.eucl <- dist(hc_df_scaled, method = "euclidean")

hc_cluster <- hclust(dist.eucl, method = "ward.D2") # Ward's method
hc_cluster


set_cutOff <- 20


plot(hc_cluster)
abline(h = set_cutOff, col = "red", lty = 2)
rect.hclust(hc_cluster, h=set_cutOff, border = "tomato")

groups <- cutree(hc_cluster, h=set_cutOff)
# groupsOut <- names(table(groups))[table(groups) >= 2]

table(groups)
aggregate(hc_df, by=list(cluster=groups), mean, na.rm = TRUE)  
```




## Latent profile analysis
just a preliminary analysis to get a feeling for what we've got.

```{r}
# lpa <- valenceDataShort %>% select(-CAM) %>%
#  estimate_profiles(1:9, variances = c("equal", "equal", "varying"),
#                         covariances = c("zero", "equal", "zero"))
```
